+++ ext/log-collector/src/main/java/core/log/LogCollectorApp.java
@@ -33,13 +33,23 @@ public class LogCollectorApp extends App {
         bind(SendEventRequestValidator.class);
 
         List<String> allowedOrigins = allowedOrigins(requiredProperty("app.allowedOrigins"));
-        EventController controller = bind(new EventController(allowedOrigins));
+        List<String> collectCookies = collectCookies(property("app.cookies").orElse(null));
+        EventController controller = bind(new EventController(allowedOrigins, collectCookies));
         http().route(HTTPMethod.OPTIONS, "/event/:app", controller::options);
         http().route(HTTPMethod.POST, "/event/:app", controller::post);  // event will be sent via ajax or navigator.sendBeacon(), refer to https://developer.mozilla.org/en-US/docs/Web/API/Navigator/sendBeacon
         Bean.register(SendEventRequest.class);
     }
 
+    List<String> collectCookies(String value) {
+        if (value == null) return null;
+        return parseList(value);
+    }
+
     List<String> allowedOrigins(String value) {
+        return parseList(value);
+    }
+
+    private List<String> parseList(String value) {
         String[] origins = Strings.split(value, ',');
         return Arrays.stream(origins).map(String::strip).collect(Collectors.toList());
     }