+++ ext/log-collector/src/main/java/core/log/web/Cookie.java
@@ -1,0 +1,14 @@
+package core.log.web;
+
+/**
+ * @author neo
+ */
+class Cookie {
+    final String name;
+    final String value;
+
+    Cookie(String name, String value) {
+        this.name = name;
+        this.value = value;
+    }
+}