+++ ext/log-collector/src/main/java/core/log/web/EventController.java
@@ -8,6 +8,7 @@ import core.framework.json.Bean;
 import core.framework.json.JSON;
 import core.framework.kafka.MessagePublisher;
 import core.framework.log.message.EventMessage;
+import core.framework.web.CookieSpec;
 import core.framework.web.Request;
 import core.framework.web.Response;
 import core.framework.web.exception.BadRequestException;
@@ -17,12 +18,14 @@ import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author neo
  */
 public class EventController {
     private final List<String> allowedOrigins;
+    private final List<String> collectCookies;
 
     @Inject
     MessagePublisher<EventMessage> eventMessagePublisher;
@@ -30,7 +33,7 @@ public class EventController {
     SendEventRequestValidator validator;
     private boolean allowAllOrigins;
 
-    public EventController(List<String> allowedOrigins) {
+    public EventController(List<String> allowedOrigins, List<String> collectCookies) {
         this.allowedOrigins = new ArrayList<>(allowedOrigins.size());
         for (String origin : allowedOrigins) {
             if ("*".equals(origin)) {
@@ -39,6 +42,7 @@ public class EventController {
                 this.allowedOrigins.add(origin);
             }
         }
+        this.collectCookies = collectCookies;
     }
 
     public Response options(Request request) {
@@ -60,21 +64,41 @@ public class EventController {
         String userAgent = request.header(HTTPHeaders.USER_AGENT).orElse(null);
         Instant now = Instant.now();
         String clientIP = request.clientIP();
+        List<Cookie> cookies = cookies(request);
 
         SendEventRequest eventRequest = sendEventRequest(request);
 
         for (SendEventRequest.Event event : eventRequest.events) {
             EventMessage message = message(event, app, now);
-
-            if (userAgent != null) message.context.put("user_agent", userAgent);
-            message.context.put("client_ip", clientIP);
-
+            addContext(message.context, userAgent, cookies, clientIP);
             eventMessagePublisher.publish(message.id, message);
         }
 
         Response response = Response.empty();
         if (origin != null) response.header("Access-Control-Allow-Origin", origin);
         return response;
+    }
+
+    void addContext(Map<String, String> context, String userAgent, List<Cookie> cookies, String clientIP) {
+        if (userAgent != null) context.put("user_agent", userAgent);
+        context.put("client_ip", clientIP);
+
+        if (cookies != null) {
+            for (Cookie cookie : cookies) {
+                context.put(cookie.name, cookie.value);
+            }
+        }
+    }
+
+    List<Cookie> cookies(Request request) {
+        if (collectCookies == null) return null;
+
+        List<Cookie> cookies = new ArrayList<>(collectCookies.size());
+        for (String name : collectCookies) {
+            request.cookie(new CookieSpec(name))
+                   .ifPresent(value -> cookies.add(new Cookie(name, value)));
+        }
+        return cookies;
     }
 
     // ignore content-type and assume it's json, due to client side may uses "navigator.sendBeacon(url, json);" to send event