+++ ext/log-collector/src/test/java/core/log/LogCollectorAppTest.java
@@ -23,4 +23,12 @@ class LogCollectorAppTest {
         List<String> origins = app.allowedOrigins("origin1, \norigin2  ,\n\rorigin3");
         assertThat(origins).containsExactly("origin1", "origin2", "origin3");
     }
+
+    @Test
+    void collectCookies() {
+        assertThat(app.collectCookies(null)).isNull();
+
+        List<String> cookies = app.collectCookies("visitor_id");
+        assertThat(cookies).containsExactly("visitor_id");
+    }
 }