+++ ext/log-collector/src/test/java/core/log/web/EventControllerTest.java
@@ -4,6 +4,7 @@ import core.framework.inject.Inject;
 import core.framework.json.Bean;
 import core.framework.log.message.EventMessage;
 import core.framework.util.Strings;
+import core.framework.web.CookieSpec;
 import core.framework.web.Request;
 import core.framework.web.exception.ForbiddenException;
 import core.log.IntegrationTest;
@@ -13,7 +14,9 @@ import org.junit.jupiter.api.Test;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static core.log.web.SendEventRequest.Event;
@@ -38,13 +41,13 @@ class EventControllerTest extends IntegrationTest {
 
     @Test
     void checkOriginWithWildcard() {
-        var controller = new EventController(List.of("*"));
+        var controller = new EventController(List.of("*"), null);
         controller.checkOrigin("https://localhost");
     }
 
     @Test
     void checkOrigin() {
-        var controller = new EventController(List.of("localhost", "example.com"));
+        var controller = new EventController(List.of("localhost", "example.com"), null);
         controller.checkOrigin("https://localhost");
         controller.checkOrigin("https://api.example.com");
 
@@ -56,7 +59,7 @@ class EventControllerTest extends IntegrationTest {
     @Test
     void options() {
         when(request.header("Origin")).thenReturn(Optional.empty());
-        var controller = new EventController(List.of("*"));
+        var controller = new EventController(List.of("*"), null);
         assertThatThrownBy(() -> controller.options(request))
                 .isInstanceOf(ForbiddenException.class);
     }
@@ -64,7 +67,7 @@ class EventControllerTest extends IntegrationTest {
     @Test
     void allowPOSTWithCORS() {
         when(request.header("Origin")).thenReturn(Optional.of("localhost"));
-        var controller = new EventController(List.of("*"));
+        var controller = new EventController(List.of("*"), null);
         assertThat(controller.options(request).header("Access-Control-Allow-Methods"))
                 .hasValueSatisfying(methods -> assertThat(methods).contains("POST"));
     }
@@ -80,7 +83,7 @@ class EventControllerTest extends IntegrationTest {
         event.info.put("stackTrace", "trace");
         event.elapsedTime = 100L;
 
-        var controller = new EventController(List.of());
+        var controller = new EventController(List.of(), null);
         Instant now = event.date.plusHours(1).toInstant();
         EventMessage message = controller.message(event, "test", now);
 
@@ -109,10 +112,35 @@ class EventControllerTest extends IntegrationTest {
 
         when(request.body()).thenReturn(Optional.of(Strings.bytes(Bean.toJSON(sendEventRequest))));
 
-        var controller = new EventController(List.of());
+        var controller = new EventController(List.of(), null);
         controller.validator = validator;
         SendEventRequest parsedSendEventRequest = controller.sendEventRequest(request);
 
         assertThat(parsedSendEventRequest).usingRecursiveComparison().isEqualTo(sendEventRequest);
+    }
+
+    @Test
+    void cookies() {
+        var controller = new EventController(List.of(), List.of("visitor_id"));
+
+        when(request.cookie(new CookieSpec("visitor_id"))).thenReturn(Optional.empty());
+        List<Cookie> cookies = controller.cookies(request);
+        assertThat(cookies).hasSize(0);
+
+        when(request.cookie(new CookieSpec("visitor_id"))).thenReturn(Optional.of("value"));
+        cookies = controller.cookies(request);
+        assertThat(cookies).hasSize(1);
+        Cookie cookie = cookies.get(0);
+        assertThat(cookie.name).isEqualTo("visitor_id");
+        assertThat(cookie.value).isEqualTo("value");
+    }
+
+    @Test
+    void addContext() {
+        var controller = new EventController(List.of(), null);
+        Map<String, String> context = new HashMap<>();
+        controller.addContext(context, "agent", List.of(new Cookie("cookie", "value")), "ip");
+
+        assertThat(context).containsKeys("client_ip", "user_agent", "cookie");
     }
 }