+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 7.3.12 (3/25/2020 - 
+* db: added query.in(field, params) shortcut to build dynamic "where in clause"
 
 ### 7.3.11 (3/23/2020 - 3/25/2020)
 * kafka: set linger.ms to 5ms by default to improve batching under load