+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Change log
+### 7.3.13 (3/31/2020 - 
+
 ### 7.3.12 (3/25/2020 - 3/30/2020)
 * db: added query.in(field, params) shortcut to build dynamic "where in clause"
 * log-collector: added param app.cookies to collect specific cookies from request. (this is due to safari expires localStorage and cookies write by js, only way to long term track is thru first party cookie)